replayer/pyproject.toml
[project]
name = "replayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "redis>=7.1.0",
]
